triangle area calculations. Is trinangle is rectangular

+++ MindboxTaskLib/MindboxTaskLib/IShapeTool.cs
@@ -8,6 +8,6 @@ namespace MindboxTaskLib
         bool IsTriangleIsRectangular(Triangle triangle);
 
         /// <exception cref="CalculatorNotFoundException"></exception>
-        decimal CalculateSquare(IShape shape);
+        double CalculateSquare(IShape shape);
     }
 }
